// src/generated/api/orders/models/posOrderItemDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export interface PosOrderItemDto {
  /** @pattern ^[1-9][0-9]*$ */
  productVariantId: string;
  /** @minimum 1 */
  quantity: number;
}